# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/reviewgame": {
            "get": {
                "description": "Get a list of ReviewGame.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReviewGame"
                ],
                "summary": "Get all ReviewGame.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Review"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new ReviewGame.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReviewGame"
                ],
                "summary": "Create New ReviewGame.",
                "parameters": [
                    {
                        "description": "the body to create a new reviewgame",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    }
                }
            }
        },
        "/reviewgame/{id}": {
            "get": {
                "description": "Get a ReviewGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReviewGame"
                ],
                "summary": "Get ReviewGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reviewgame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a ReviewGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReviewGame"
                ],
                "summary": "Delete one ReviewGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reviewgame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update ReviewGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReviewGame"
                ],
                "summary": "Update ReviewGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reviewgame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an ReviewGame",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.reviewInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.reviewInput": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "game_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Review": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}